Update the bmi_calc.c
This commit added the neccessary conditional statements to reflect the
approprate feedback messages of a BMI calculator. Before this commit,
there was just a basic BMI mathematical layout which served as a bulding
block base. I used relational and logical operators in the conditional
statements because it will be easier to maintain.